user instructions

+++ ToDoCalender/Presentation Layer/ToDoCalender/MoreAboutTask.cs
@@ -14,15 +14,15 @@ namespace ToDoCalender
 {
     public partial class MoreAboutTask : Form
     {
-        private ToDoCalenderWindow OriginalWindow {  get; set; }
-        public TaskToDo TheTask {  get; set; }
+        private ToDoCalenderWindow OriginalWindow { get; set; }
+        public TaskToDo TheTask { get; set; }
         public MoreAboutTask(TaskToDo aTask, object sender)
         {
             TheTask = aTask;
             InitializeComponent();
             setComponentValues();
             OriginalWindow = (ToDoCalenderWindow)sender;
-     
+
         }
 
         public void setComponentValues()
@@ -34,8 +34,16 @@ namespace ToDoCalender
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            OriginalWindow.DeleteTask(TheTask);
-            this.Close();
+            if (InputValidation.ValidateDelete())
+            {
+                MessageBox.Show("Task Deleted!");
+                OriginalWindow.DeleteTask(TheTask);
+                this.Close();
+            }
+            else
+            {
+                MessageBox.Show("Task was not deleted!");
+            }
         }
     }
 }
